requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.0.0
apispec==5.2.2
apispec-webframeworks==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Automat==22.10.0
Babel==2.13.1
backcall==0.2.0
bcc==0.18.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
bidict==0.22.1
binaryornot==0.4.4
bleach==6.1.0
blinker==1.7.0
Brlapi==0.8.3
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.1
chrome-gnome-shell==0.0.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.4
comm==0.1.4
command-not-found==0.3
configobj==5.0.6
constantly==23.10.4
contourpy==1.1.1
cookiecutter==2.1.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dask==2024.1.1
dataclasses-json==0.6.3
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
deepdiff==6.2.3
defer==1.0.6
defusedxml==0.7.1
discord==2.3.2
discord.py==2.3.2
distributed==2024.1.1
distro==1.7.0
dnspython==2.5.0
duplicity==0.8.21
ecdsa==0.18.0
et-xmlfile==1.1.0
evdev==1.4.0
exceptiongroup==1.1.3
executing==2.0.0
fasteners==0.14.1
fastjsonschema==2.18.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.12.2
future==0.18.2
gast==0.5.4
gevent==23.9.1
gevent-websocket==0.10.1
gitignore-parser==0.1.11
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.60.0
gyp==0.1
h11==0.14.0
h5py==3.10.0
hidpidaemon==18.4.6
html5lib==1.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==7.0.1
incremental==22.10.0
ipykernel==6.26.0
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
jinja2-time==0.2.0
json5==0.9.14
jsonify==0.5
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.4.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kazam==1.4.5
keras==2.15.0
kernelstub==3.1.4
keyboard==0.13.5
keyring==23.5.0
kiwisolver==1.4.5
kthread==0.2.3
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.16
langsmith==0.0.83
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.6
locket==1.0.0
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
marshmallow-sqlalchemy==0.28.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
modin==0.26.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
netaddr==0.8.0
netifaces==0.11.0
networkx==2.8.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.1
oauthlib==3.2.0
olefile==0.46
openai==0.28.1
openpyxl==3.1.2
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.9.12
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
panda==0.3.1
pandas==2.1.1
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
partd==1.4.1
passlib==1.7.4
pbkdf2==1.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
pop-transition==1.1.2
powerline-status==2.7
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.5.3
pydantic_core==2.14.6
pydbus==0.6.0
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.6.1
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-engineio==4.8.2
python-gnupg==0.4.8
python-json-logger==2.0.7
python-linux-procfs==0.6.3
python-magic==0.4.24
python-slugify==8.0.2
python-socketio==5.11.0
python-xlib==0.29
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyudev==0.22.0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.1
redis==5.0.1
referencing==0.30.2
repolib==2.2.1
repoman==1.4.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.10.6
rsa==4.9
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
selenium==4.17.2
Send2Trash==1.8.2
sessioninstaller==0.0.0
setproctitle==1.2.2
simple-websocket==0.10.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.1
spotipy==2.23.0
SQLAlchemy==2.0.25
srp==1.0.20
ssh-import-id==5.11
stack-data==0.6.3
sympy==1.12
systemd-python==234
taipy==3.0.0
taipy-config==3.0.0
taipy-core==3.0.0
taipy-gui==3.0.0
taipy-rest==3.0.0
taipy-templates==3.0.0
tblib==3.0.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
terminado==0.17.1
text-unidecode==1.3
tinycss2==1.2.1
tlslite-ng==0.8.0a43
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
trio==0.24.0
trio-websocket==0.11.1
Twisted==23.10.0
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==4.3.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unicorn==2.0.1.post1
uri-template==1.3.0
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
yarl==1.9.2
zict==3.0.0
zipp==1.0.0
zope.event==5.0
zope.interface==6.1
